LibWeb: Move Declaration code into Declaration.cpp
