+++ Userland/Libraries/LibWeb/CMakeLists.txt
@@ -46,6 +46,7 @@ set(SOURCES
     CSS/MediaQueryList.cpp
     CSS/Parser/Block.cpp
     CSS/Parser/ComponentValue.cpp
+    CSS/Parser/Declaration.cpp
     CSS/Parser/Function.cpp
     CSS/Parser/Parser.cpp
     CSS/Parser/StyleRules.cpp
